custom/requirements.txt
jupyterhub-kubespawner
oauthenticator